==> 00 - Proceso batch total de todos los programas.py <==
import runpy

try:
    runpy.run_path('1 - Bajo los excel.py')
except Exception as e:
    print(f"Error al ejecutar 1 - Bajo los excel.py: {e}")

for i in range(0,5):
    print('######################')

try:
    runpy.run_path('2 - Armado de los txt.py')
except Exception as e:
    print(f"Error al ejecutar 2 - Armado de los txt_p.py: {e}")

for i in range(0,5):
    print('######################')

try:
    runpy.run_path('3 - Cuadro de cotizaciones x mes total.py')
except Exception as e:
    print(f"Error al ejecutar 3 - Cuadro de cotizaciones x mes total.py: {e}")

for i in range(0,5):
    print('######################')

try:
    runpy.run_path('3_Uno los archivos Enlaces y Parrafos por separado.py')
except Exception as e:
    print(f"Error al ejecutar 3_Uno los archivos Enlaces y Parrafos por separado.py: {e}")

for i in range(0,5):
    print('######################')

try:
    runpy.run_path('4 - Armo el archivo para transferir a Wordpress sql.py')
except Exception as e:
    print(f"Error al ejecutar 4 - Armo el archivo para transferir a Wordpress sql.py: {e}")

for i in range(0,5):
    print('######################')

try:
    runpy.run_path('5 - Genero el sql.py')
except Exception as e:
    print(f"Error al ejecutar 5 - Genero el sql.py: {e}")

for i in range(0,5):
    print('######################')

try:
    runpy.run_path('6 - Genero los archivos excel.py')
except Exception as e:
    print(f"Error al ejecutar 6 - Genero los archivos excel.py: {e}")

for i in range(0,5):
    print('######################')

for i in range(0,5):
    print('######################')
for i in range(0,5):
    print('######################')
print('T E R M I N A D O')
